Move ObjectKind to transformation to avoid circular dependencies between model and transformation
