+++ ts/libs/core/src/lib/transformation/public/common/index.ts
@@ -1,4 +1,5 @@
 export * from './errors';
+export * from './object-kind.prm';
 export * from './polaris-transformation-config';
 export * from './polaris-transformation-service-manager';
 export * from './polaris-transformation-service';
